Make necessary metrics API symbols private
Fixes #2622
