+++ docs/api/metrics.instrument.rst
@@ -1,7 +1,7 @@
 opentelemetry._metrics.instrument
 =================================
 
-.. automodule:: opentelemetry._metrics.instrument
+.. automodule:: opentelemetry._metrics._instrument
     :members:
     :private-members:
     :undoc-members:
