+++ docs/api/metrics.observation.rst
@@ -1,7 +1,7 @@
 opentelemetry._metrics.observation
 ==================================
 
-.. automodule:: opentelemetry._metrics.observation
+.. automodule:: opentelemetry._metrics._observation
     :members:
     :undoc-members:
     :show-inheritance:
