+++ docs/api/metrics.rst
@@ -3,7 +3,7 @@ opentelemetry._metrics package
 
 .. warning::
    OpenTelemetry Python metrics are in an experimental state. The APIs within
-   :mod:`opentelemetry._metrics` are subject to change in minor/patch releases and make no
+   :mod:`opentelemetry._metrics.___init__` are subject to change in minor/patch releases and make no
    backward compatability guarantees at this time.
 
    Once metrics become stable, this package will be be renamed to ``opentelemetry.metrics``.
@@ -19,4 +19,4 @@ Submodules
 Module contents
 ---------------
 
-.. automodule:: opentelemetry._metrics
+.. automodule:: opentelemetry._metrics.___init__
