+++ opentelemetry-api/src/opentelemetry/_metrics/observation.py
@@ -12,41 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from typing import Union
 
-from opentelemetry.util.types import Attributes
+from opentelemetry._metrics._observation import Observation
 
-
-class Observation:
-    """A measurement observed in an asynchronous instrument
-
-    Return/yield instances of this class from asynchronous instrument callbacks.
-
-    Args:
-        value: The float or int measured value
-        attributes: The measurement's attributes
-    """
-
-    def __init__(
-        self, value: Union[int, float], attributes: Attributes = None
-    ) -> None:
-        self._value = value
-        self._attributes = attributes
-
-    @property
-    def value(self) -> Union[float, int]:
-        return self._value
-
-    @property
-    def attributes(self) -> Attributes:
-        return self._attributes
-
-    def __eq__(self, other: object) -> bool:
-        return (
-            isinstance(other, Observation)
-            and self.value == other.value
-            and self.attributes == other.attributes
-        )
-
-    def __repr__(self) -> str:
-        return f"Observation(value={self.value}, attributes={self.attributes})"
+__all__ = ["Observation"]
