+++ tests/opentelemetry-test-utils/src/opentelemetry/test/globals_test.py
@@ -14,8 +14,8 @@
 
 import unittest
 
-from opentelemetry import _metrics as metrics_api
 from opentelemetry import trace as trace_api
+from opentelemetry._metrics import ___init__ as metrics_api
 from opentelemetry.util._once import Once
